pyproject.toml
[project]
name = "studyflow-ai"
version = "2.10.0"
description = "AI-powered local-first study workspace for PDFs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.1",
  "chromadb>=0.5.5",
  "easyocr>=1.7.1",
  "fastapi>=0.111.0",
  "httpx>=0.27.0",
  "pymupdf>=1.24.5",
  "pytesseract>=0.3.10",
  "rank-bm25>=0.2.2",
  "requests>=2.31.0",
  "sentence-transformers>=2.7.0",
  "streamlit>=1.32.0",
  "streamlit-autorefresh>=1.0.1",
  "typer>=0.12.3",
  "uvicorn>=0.30.0",
  "pytest>=8.2.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
studyflow = "cli.main:app"

[tool.setuptools]
packages = [
  "app",
  "app.components",
  "app.ui",
  "app.adapters",
  "cli",
  "cli.commands",
  "backend",
  "core",
  "core.assets",
  "core.coach",
  "core.config",
  "core.external",
  "core.concepts",
  "core.related",
  "core.bundle",
  "core.packs",
  "core.tasks",
  "core.quality",
  "core.agents",
  "core.formatting",
  "core.ingest",
  "core.indexing",
  "core.ui_state",
  "core.parsing",
  "core.prompts",
  "core.plugins",
  "core.plugins.builtins",
  "core.retrieval",
  "core.telemetry",
  "core.llm",
  "core.llm.providers",
  "infra",
  "service",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4.0",
  "pytest>=8.2.0",
]

[tool.ruff]
target-version = "py310"
line-length = 120
exclude = [
  ".git",
  ".venv",
  "__pycache__",
  "*.egg-info",
  ".pytest_cache",
  ".ruff_cache",
  "workspaces",
]

[tool.ruff.lint]
select = [
  "E",      # pycodestyle errors
  "W",      # pycodestyle warnings
  "F",      # pyflakes
  "I",      # isort
  "B",      # flake8-bugbear
  "UP",     # pyupgrade
]
ignore = [
  "E501",   # line too long (handled by formatter)
  "B008",   # do not perform function calls in argument defaults
  "B904",   # raise from
  "B905",   # zip without explicit strict
  "E402",   # module level import not at top of file
]

[tool.ruff.lint.isort]
known-first-party = ["app", "cli", "core", "infra", "service", "backend"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q --tb=short"
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
